+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The formatter is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Use singular noun in summary when count is 1 ([#100](https://github.com/cucumber/pretty-formatter/pull/100))
 
 ## [3.3.0] - 2026-05-03
 ### Added